refactor(api): Break game-config into config component
