+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export { InstanceOptions } from "./instance-options";
+export { GameOptions, DEFAULT_GAME_OPTIONS } from "./game-options";
+export { GameMetadata } from "./game-metadata";
